Create a helper static class to avoid code duplication
