+++ src/PublicApiAnalyzers/Core/CodeFixes/AnnotatePublicApiFix.cs
@@ -41,7 +41,7 @@ namespace Microsoft.CodeAnalysis.PublicApiAnalyzers
                 string publicSymbolNameWithNullability = diagnostic.Properties[DeclarePublicApiAnalyzer.PublicApiNameWithNullabilityPropertyBagKey];
                 bool isShippedDocument = diagnostic.Properties[DeclarePublicApiAnalyzer.PublicApiIsShippedPropertyBagKey] == "true";
 
-                TextDocument? document = isShippedDocument ? DeclarePublicApiFix.GetShippedDocument(project) : DeclarePublicApiFix.GetUnshippedDocument(project);
+                TextDocument? document = isShippedDocument ? PublicApiFixHelpers.GetShippedDocument(project) : PublicApiFixHelpers.GetUnshippedDocument(project);
 
                 if (document != null)
                 {
@@ -82,14 +82,8 @@ namespace Microsoft.CodeAnalysis.PublicApiAnalyzers
                 }
             }
 
-            var endOfLine = Environment.NewLine;
-            if (sourceText.Lines.Count > 1)
-            {
-                var firstLine = sourceText.Lines[0];
-                endOfLine = sourceText.ToString(new TextSpan(firstLine.End, firstLine.EndIncludingLineBreak - firstLine.End));
-            }
-
-            SourceText newSourceText = sourceText.Replace(new TextSpan(0, sourceText.Length), string.Join(endOfLine, lines) + DeclarePublicApiFix.GetEndOfFileText(sourceText, endOfLine));
+            var endOfLine = PublicApiFixHelpers.GetEndOfLine(sourceText);
+            SourceText newSourceText = sourceText.Replace(new TextSpan(0, sourceText.Length), string.Join(endOfLine, lines) + PublicApiFixHelpers.GetEndOfFileText(sourceText, endOfLine));
             return newSourceText;
         }
 
@@ -119,7 +113,7 @@ namespace Microsoft.CodeAnalysis.PublicApiAnalyzers
                     Project project = pair.Key;
                     ImmutableArray<Diagnostic> diagnostics = pair.Value;
 
-                    TextDocument? unshippedDocument = DeclarePublicApiFix.GetUnshippedDocument(project);
+                    TextDocument? unshippedDocument = PublicApiFixHelpers.GetUnshippedDocument(project);
                     if (unshippedDocument?.FilePath != null && !uniqueUnshippedDocuments.Add(unshippedDocument.FilePath))
                     {
                         // Skip past duplicate unshipped documents.
@@ -127,7 +121,7 @@ namespace Microsoft.CodeAnalysis.PublicApiAnalyzers
                         unshippedDocument = null;
                     }
 
-                    TextDocument? shippedDocument = DeclarePublicApiFix.GetShippedDocument(project);
+                    TextDocument? shippedDocument = PublicApiFixHelpers.GetShippedDocument(project);
                     if (shippedDocument?.FilePath != null && !uniqueShippedDocuments.Add(shippedDocument.FilePath))
                     {
                         // Skip past duplicate shipped documents.
